+++ example/App.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { ScrollView, Text, View } from 'react-native';
+import { ScrollView, Text, View, SafeAreaView } from 'react-native';
 import { MultiSelect, Select, type Data } from '@rose-hulman/react-native-dropdown-selector';
+import { useThemeStyles } from './styles';
 
 const data: Data[] = [
   { label: 'Item 1' },
@@ -25,10 +26,24 @@ const themes: Data[] = [
 ];
 
 function App(): React.JSX.Element {
+  const [theme, setTheme] = React.useState<'light' | 'dark' | 'system'>('system');
+  const onThemeSelect = (datum: Data) => {
+    setTheme(datum.label as 'light' | 'dark' | 'system');
+  }
+
+  return (
+    <Content
+      onThemeSelect={onThemeSelect}
+      theme={theme}
+    />);
+}
+
+const Content = ({ onThemeSelect, theme }: ContentProperties): React.JSX.Element => {
   const [item, setItem] = React.useState<string | JSX.Element>('');
   const [disabled, setDisabled] = React.useState(false);
   const [searchable, setSearchable] = React.useState(false);
-  const [theme, setTheme] = React.useState<'light' | 'dark' | 'system'>('system');
+  const style = useThemeStyles(theme);
+
   const onSimpleDataSelect = (datum: Data) =>
     setItem(datum.label);
   const onSimpleMultiDataSelect = (data: Data[]) =>
@@ -44,12 +59,9 @@ function App(): React.JSX.Element {
   const onMultiDataSelect = (id: number) => (data: Data[]) => {
     console.log(`selector ${id}: currently contains ${data.map(datum => datum.label).join(", ")}.`);
   };
-  const onThemeSelect = (datum: Data) =>
-    setTheme(datum.label as 'light' | 'dark' | 'system');
 
   return (
-    <>
-      <View style={{ height: 40 }} />
+    <SafeAreaView style={style.background}>
       <ScrollView style={{ paddingHorizontal: 8 }}>
         <View style={{ height: 40 }} />
         <Select
@@ -59,13 +71,17 @@ function App(): React.JSX.Element {
           searchable={searchable}
           theme={theme}
         />
-        <Text>Selected: {item || 'None'} (scroll down)</Text>
+        <Text style={style.text}>
+          Selected: {item || 'None'} (scroll down)
+        </Text>
         <View style={{ height: 500 }} />
         <Text
-          style={{
+          style={[
+            style.text,
+            {
             alignSelf: 'center',
             width: 350,
-          }}
+          }]}
         >
           The dropdown menu will display above the input box when there
           isn&apos;t enough space below
@@ -90,7 +106,7 @@ function App(): React.JSX.Element {
           theme={theme}
         />
         <View style={{ height: 400 }}/>
-        <Text>Single Selects:</Text>
+        <Text style={style.text}>Single Selects:</Text>
         <View style={{ flexDirection: 'row', height: 100 }}>
           <View style={{ flex: 1 }}>
             <Select
@@ -120,7 +136,7 @@ function App(): React.JSX.Element {
             />
           </View>
         </View>
-        <Text>Multi Selects:</Text>
+        <Text style={style.text}>Multi Selects:</Text>
         <View style={{ height: 350 }}>
           <View style={{ flexDirection: 'row' }}>
             <View style={{ flex: 1 }}>
@@ -154,7 +170,7 @@ function App(): React.JSX.Element {
               />
             </View>
           </View>
-          <Text style={{ textAlign: 'center', height: 100 }}>
+          <Text style={[style.text, { textAlign: 'center', height: 100 }]}>
             Select more than one item and see me move!
           </Text>
           <MultiSelect
@@ -187,11 +203,11 @@ function App(): React.JSX.Element {
             placeholderText={`Select a theme`}
             theme={theme}
           />
-          <Text style={{ textAlign: 'center'}}>
+          <Text style={[style.text, { textAlign: 'center'}]}>
             Select a theme to see all the dropdowns change! Current theme is "{theme}"
           </Text>
         </View>
-        <Text>Styled Single Select:</Text>
+        <Text style={style.text}>Styled Single Select:</Text>
         <Select
           data={data}
           onSelect={onSimpleDataSelect}
@@ -270,7 +286,7 @@ function App(): React.JSX.Element {
           }}
         />
         <View style={{height: 50}} />
-        <Text>Styled Multi Select:</Text>
+        <Text style={style.text}>Styled Multi Select:</Text>
         <MultiSelect
           data={data}
           onSelect={onSimpleMultiDataSelect}
@@ -370,8 +386,13 @@ function App(): React.JSX.Element {
         />
         <View style={{ height: 700 }} />
       </ScrollView>
-    </>
+    </SafeAreaView>
   );
+}
+
+interface ContentProperties {
+  onThemeSelect: (e: Data) => void;
+  theme: 'light' | 'dark' | 'system';
 }
 
 export default App;